use provider for first post published modal
