+++ apps/editing-toolkit/editing-toolkit-plugin/wpcom-block-editor-nux/src/block-editor-nux.js
@@ -10,6 +10,7 @@ import { useEffect, useState } from '@wordpress/element';
 import { registerPlugin } from '@wordpress/plugins';
 import { __dangerousOptInToUnstableAPIsOnlyForCoreModules } from '@wordpress/private-apis';
 import { getQueryArg } from '@wordpress/url';
+import { ShouldShowFirstPostPublishedModalProvider } from '../../dotcom-fse/lib/first-post-published-modal/should-show-first-post-published-modal-context';
 import { HasSeenSellerCelebrationModalProvider } from '../../dotcom-fse/lib/seller-celebration-modal/has-seen-seller-celebration-modal-context';
 import { HasSeenVideoCelebrationModalProvider } from '../../dotcom-fse/lib/video-celebration-modal/has-seen-video-celebration-modal-context';
 import DraftPostModal from './draft-post-modal';
@@ -101,11 +102,13 @@ registerPlugin( 'wpcom-block-editor-nux', {
 	render: () => (
 		<HasSeenSellerCelebrationModalProvider>
 			<HasSeenVideoCelebrationModalProvider>
-				<WelcomeTour />
-				<PostPublishedModal />
-				<SellerCelebrationModal />
-				<PurchaseNotice />
-				<VideoPressCelebrationModal />
+				<ShouldShowFirstPostPublishedModalProvider>
+					<WelcomeTour />
+					<PostPublishedModal />
+					<SellerCelebrationModal />
+					<PurchaseNotice />
+					<VideoPressCelebrationModal />
+				</ShouldShowFirstPostPublishedModalProvider>
 			</HasSeenVideoCelebrationModalProvider>
 		</HasSeenSellerCelebrationModalProvider>
 	),
