+++ apps/editing-toolkit/editing-toolkit-plugin/wpcom-block-editor-nux/src/post-published-modal/index.tsx
@@ -5,7 +5,7 @@ import { useEffect, useRef, useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { isURL } from '@wordpress/url';
 import React from 'react';
-import useShouldShowFirstPostPublishedModal from '../../../dotcom-fse/lib/first-post-published-modal/use-should-show-first-post-published-modal';
+import { useShouldShowFirstPostPublishedModal } from '../../../dotcom-fse/lib/first-post-published-modal/should-show-first-post-published-modal-context';
 import NuxModal from '../nux-modal';
 import postPublishedImage from './images/post-published.svg';
